tugger-apple-codesign: serialize length value as u32
We were writing this out as usize, which was likely 64-bits.
This was causing a roundtripping failure.

After this commit, we can now roundtrip all blobs from all Mach-O
binaries on one of my macOS machines!

+++ tugger-apple-codesign/src/macho.rs
@@ -1609,7 +1609,7 @@ impl<'a> Blob<'a> for OtherBlob<'a> {
     fn to_vec(&self) -> Result<Vec<u8>, MachOError> {
         let mut res = Vec::with_capacity(self.data.len() + 8);
         res.iowrite_with(self.magic, scroll::BE)?;
-        res.iowrite_with(self.data.len() + 8, scroll::BE)?;
+        res.iowrite_with(self.data.len() as u32 + 8, scroll::BE)?;
         res.write_all(&self.data)?;
 
         Ok(res)
